eatgo-api/src/main/java/me/restaurant/domain/Restaurant.java
package me.restaurant.domain;

public class Restaurant {
    private final String name;
    private final String address;
    private final Long id;

    public Restaurant(Long id, String name, String address) {
        this.id = id;
        this.name = name;
        this.address = address;
    }

    public Long getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public String getAddress() {
        return address;
    }

    public String information() {
        return name + " in " + address;
    }
}
